MAUI Classic Document: - Added Document Scanner Classic Component full example with Binding properties. - Moved the Scanners in corresponding folders - Added a home page for scanner listing - Deleted the previous MainPage
