+++ MAUI/ClassicComponent.Maui/App.xaml.cs
@@ -7,7 +7,7 @@
         public App()
         {
             InitializeComponent();
-            MainPage = new MainPage();
+            MainPage = new NavigationPage(new HomePage());
         }
     }
 }
