+++ MAUI/ClassicComponent.Maui/Pages/BarcodeScanner/ClassicBarcodeScannerPage.xaml
@@ -1,7 +1,7 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
+<?xml version="1.0" encoding="utf-8"?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="ClassicComponent.Maui.MainPage"
+             x:Class="ClassicComponent.Maui.ClassicBarcodeScannerPage"
              xmlns:ios="clr-namespace:Microsoft.Maui.Controls.PlatformConfiguration.iOSSpecific;assembly=Microsoft.Maui.Controls"
              xmlns:local="clr-namespace:ClassicComponent.Maui.CustomViews"
              ios:Page.UseSafeArea="true"
