--- MAUI/ClassicComponent.Maui/Pages/ClassicDocumentScannerPage.xaml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="ClassicComponent.Maui.ClassicDocumentScannerPage"
-             xmlns:classicView="clr-namespace:ScanbotSDK.MAUI.ClassicComponents;assembly=DocumentSDK.MAUI">
-    <ContentPage.Content>
-        <Grid RowDefinitions="2*,7*,1*">
-            <Label
-                Grid.Row="0"
-                x:Name="labelHeading"
-                FontSize="30"
-                HorizontalTextAlignment="Center"
-                VerticalTextAlignment="Center"
-                Text="Classic Components Yay!!!"
-                BackgroundColor="DarkSlateBlue" />
-            <classicView:ClassicDocumentScannerView
-                Grid.Row="1"
-                BackgroundColor="Crimson"
-                x:Name="documentScannerView"
-                IsFlashEnabled="True"
-                OnDocumentImageSnapped="DocumentScannerView_OnDocumentImageSnapped" />
-            <StackLayout
-                Grid.Row="2">
-                <HorizontalStackLayout
-                    HorizontalOptions="Center"
-                    Spacing="10">
-                    <Button
-                        Text="Start"
-                        Clicked="Button_StartCameraClicked" />
-                    <Button
-                        Text="Stop"
-                        Clicked="Button_StopCameraClicked" />
-                </HorizontalStackLayout>
-            </StackLayout>
-        </Grid>
-    </ContentPage.Content>
-</ContentPage>
