+++ MAUI/ClassicComponent.Maui/Pages/DocumentScanner/ClassicDocumentScannerPage.xaml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="ClassicComponent.Maui.ClassicDocumentScannerPage"
+             xmlns:scannerPage="clr-namespace:ClassicComponent.Maui;assembly=ClassicComponent.Maui"
+             xmlns:classicView="clr-namespace:ScanbotSDK.MAUI.ClassicComponents;assembly=DocumentSDK.MAUI">
+    <ContentPage.Content>
+        <Grid RowDefinitions="8*,2*">
+            <classicView:ClassicDocumentScannerView
+                Grid.Row="0"
+                x:Name="documentScannerView"
+                IgnoreBadAspectRatio="True"
+                AutoSnappingEnabled ="{Binding AutoSnappingEnabled}"
+                FlashEnabled="{Binding IsFlashEnabled}"
+                PolygonViewVisible="{Binding PolygonViewEnabled}" />
+            <Grid
+                Grid.Row="1"
+                BackgroundColor="DimGray">
+                    <HorizontalStackLayout
+                        HorizontalOptions="Center"
+                        VerticalOptions="Center"
+                        Spacing="10">
+                        <Button
+                            Text="{x:Static scannerPage:ClassicDocumentScannerPage.Stop}"
+                            BackgroundColor="WhiteSmoke"
+                            TextColor="SandyBrown"
+                            Padding="10"
+                            Clicked="DocumentScanner_ButtonClicked" />
+                        <Button
+                            Text="{x:Static scannerPage:ClassicDocumentScannerPage.AutoSnap}"
+                            BackgroundColor="WhiteSmoke"
+                            TextColor="SandyBrown"
+                            Padding="10" 
+                            Clicked="DocumentScanner_ButtonClicked" />
+                        <Button
+                            Text="{x:Static scannerPage:ClassicDocumentScannerPage.Flash}"
+                            BackgroundColor="WhiteSmoke"
+                            TextColor="SandyBrown"
+                            Padding="10"
+                            Clicked="DocumentScanner_ButtonClicked" />
+                        <Button
+                            Text="{x:Static scannerPage:ClassicDocumentScannerPage.Polygons}"
+                            BackgroundColor="WhiteSmoke"
+                            TextColor="SandyBrown"
+                            Padding="10"
+                            Clicked="DocumentScanner_ButtonClicked"/>
+                    </HorizontalStackLayout>
+            </Grid>
+        </Grid>
+    </ContentPage.Content>
+</ContentPage>
