+++ MAUI/ClassicComponent.Maui/Pages/DocumentScanner/DocumentScannerResultPage.xaml
